Add alarm script
Added working alarm script that prompts the user with a window
asking for the alarm time. At the specified time the alarm will go
off and turn of the smartbulb.

The IP for the smartbulb is fixed to 192.168.1.150 which needs to
be fixed.

There is no implementation for audio yet.

There is no support for changing color yet.