Add automatic setup during weekdays
If it is a evening before weekday the alarm will automatically be
set to 05:45. If Friday evening or Saturday evening the alarm will
not set timer automatically. If time is after 21:00 the alarm will
also not set up a time automatically so even during weekends it is
possible to set a custom alarm time by starting the RPi after
21:00.