Add timeout to the IP input to make the pi run headless
A timout to the input window of the bulps ip is added
to make the script run and set up the alarm without the need
for any user input so it can run in headless mode.

TODO. Timeout is set to 30 second which might be a bit to
small if the ip needs to be configured often. The ip
should be saved and restored next time the script is run.